Yaegi fails with "impossible type assertion: *backoff.backOffContext does not implement backoff.BackOff (missing Reset method)"
So we fork backoff/v3 (without generics) and fix the issue there
Update tests because maxRetries in v3 is treated differently. Now we retry 3 times so we expect to get the failure log 4 times: the original failure + 3 retries

+++ plugin.go
@@ -23,7 +23,7 @@ type Config struct {
 	Roles                        []string
 	MatchAllRoles                bool
 	UpdateKeysEveryMinutes       int
-	UpdateKeysWithBackoffRetries uint
+	UpdateKeysWithBackoffRetries uint64
 	LogLevel                     string
 	LogHeaders                   []string
 }
